Move drafter naming, ranking and DFlash discovery into a drafters package
Pure structural move, no behaviour change. The shared primitives, the
ranking keys and the DFlash detector were spread through model_config
alongside unrelated model handling, and the same rules are reached from
four different paths, so they now live in one package.

model_config re-exports every moved name, so callers and tests that
import them from there keep working.

The package deliberately does not import model_config at module import
time. The GGUF split and quant naming helpers stay where they are, since
non-drafter code shares them, and are imported per call instead.
