+++ studio/backend/utils/models/model_config.py
@@ -1656,198 +1656,27 @@ def detect_mmproj_file(path: str, search_root: Optional[str] = None) -> Optional
     return str(best[1])
 
 
-def _drafter_pairing_stem(name: str, *, kind: str) -> str:
-    """The model family a drafter filename names, stripped of its own markers.
-
-    Both published schemes are handled: ``<kind>-<model>`` and the older
-    ``<model>-<KIND>``. The shard suffix sits outside the quant token, so it
-    goes first or the anchored quant strip below cannot match. Full quant
-    vocabulary, not a subset: K/IQ/UD/MXFP drafters pair too, and the optional
-    bpw modifier goes with it, as _extract_quant_label does.
-    """
-    stem = Path(name).stem.lower()
-    if stem.startswith(f"{kind}-"):
-        stem = stem[len(kind) + 1 :]
-    stem = re.sub(r"-[0-9]{5}-of-[0-9]{5}$", "", stem)
-    if stem.endswith(f"-{kind}"):
-        stem = stem[: -(len(kind) + 1)]
-    return re.sub(
-        rf"-(?:{_GGUF_KNOWN_QUANT_RE.pattern})(?:-[0-9]+(?:\.[0-9]+)?bpw)?$",
-        "",
-        stem,
-        flags = re.IGNORECASE,
-    )
-
-
-def _drafter_matches_weight(candidate_name: str, weight_name: Optional[str], *, kind: str) -> bool:
-    """Whether a drafter pairs with the weight, by name.
-
-    A multi-model folder must not attach a foreign drafter, so the family the
-    drafter names has to PREFIX the weight filename at a non-alphanumeric
-    boundary. That blocks one direction of a ``DeepSeek-V4-Flash-Lite`` /
-    ``DeepSeek-V4-Flash`` pair but not the other: the shorter family name is a
-    prefix of the longer weight, so a base-family sidecar still matches a
-    longer-named sibling's weights. Exact equality cannot replace the prefix
-    rule -- ``mtp-gemma-4-12B-it.gguf`` really does ship beside
-    ``gemma-4-12B-it-qat-*.gguf`` -- so the remaining direction is settled by
-    ranking: callers prefer the longest matching stem (see _drafter_stem_rank).
-    """
-    if weight_name is None:
-        return True
-    stem = _drafter_pairing_stem(candidate_name, kind = kind)
-    weight = weight_name.lower()
-    return (
-        bool(stem)
-        and weight.startswith(stem)
-        and (len(weight) == len(stem) or not weight[len(stem)].isalnum())
-    )
-
-
-def _drafter_stem_rank(candidate_name: str, *, kind: str) -> int:
-    """Sort key placing the most specific family first (longest stem wins).
-
-    Both ``mtp-DeepSeek-V4-Flash-BF16.gguf`` and
-    ``mtp-DeepSeek-V4-Flash-0731-BF16.gguf`` prefix-match a 0731 weight, and
-    only the second is really its drafter.
-    """
-    return -len(_drafter_pairing_stem(candidate_name, kind = kind) or "")
-
-
-def _drafter_launch_path(candidate: Path) -> str:
-    """The path llama-server should receive for *candidate*.
-
-    llama-server takes shard 1 as the model path, and a split copy must stay on
-    its snapshot path: the blob target has no sibling shard names. Single-file
-    drafters still resolve, as callers expect.
-    """
-    loadable = _local_gguf_load_path(candidate)
-    if _GGUF_SPLIT_FILE_RE.match(loadable.name):
-        return str(loadable)
-    return str(loadable.resolve())
-
-
-def _drafter_split_is_complete(candidate: Path) -> bool:
-    """False for a partial split set, which would fail llama-server's draft
-    startup and disable speculation entirely; skip it so a complete copy wins."""
-    try:
-        _, complete = colocated_split_shards(candidate)
-    except OSError:
-        return False
-    return complete
-
-
-def _drafter_total_size(candidate: Path) -> int:
-    """Bytes across every shard. Candidates are collapsed to shard 1, so a split
-    copy must be summed or it would outrank a smaller single file."""
-    try:
-        shards, _ = colocated_split_shards(candidate)
-        return sum(shard.stat().st_size for shard in shards)
-    except OSError:
-        return sys.maxsize
-
-
-def dspark_precision_rank(name: str) -> int:
-    """Sidecar precision preference: Q8_0 first, the precision the DSpark model
-    card recommends. Shared with the hub download and VRAM-sizing paths so the
-    file Studio budgets for is the file it fetches and launches."""
-    base = Path(name).name.lower()
-    if "-q8_0" in base:
-        return 0
-    if "-q4_0" in base:
-        return 1
-    if "-bf16" in base or "-f16" in base:
-        return 2
-    return 3
-
-
-def dspark_preference_key(name: str) -> tuple[int, str]:
-    """Sort key picking the preferred sidecar by name alone (no filesystem)."""
-    return dspark_precision_rank(name), Path(name).name.lower()
-
-
-# DFlash publishes the same precision vocabulary (and the published sidecar
-# carries no precision token at all, which lands in the catch-all rank), so the
-# ordering is shared rather than duplicated.
-dflash_precision_rank = dspark_precision_rank
-
-
-def dflash_preference_key(name: str) -> tuple[int, str]:
-    """Sort key picking the preferred DFlash sidecar by name alone."""
-    return dflash_precision_rank(name), Path(name).name.lower()
-
-
-def _drafter_names_other_weight(
-    candidate_name: str,
-    weight_name: Optional[str],
-    other_weight_names: Iterable[str],
-    *,
-    kind: str = "dflash",
-) -> bool:
-    """Whether a sidecar names a DIFFERENT weight sitting beside it.
-
-    A sidecar that names no family at all (the published ``dflash-kquant.gguf``,
-    whose stem is a precision token) has to stay eligible, so "does it name a
-    family" cannot be answered from the sidecar name alone. It is answered
-    against the weights actually present instead: only a stem that pairs with
-    some OTHER weight in the same repo/folder is evidence the sidecar belongs to
-    that neighbour rather than to the weight being loaded.
-    """
-    if weight_name is None:
-        return False
-    if _drafter_matches_weight(candidate_name, weight_name, kind = kind):
-        return False
-    return any(
-        _drafter_matches_weight(candidate_name, other, kind = kind) for other in other_weight_names
-    )
-
-
-def dflash_repo_preference_key(
-    name: str,
-    weight_name: Optional[str] = None,
-    other_weight_names: Iterable[str] = (),
-) -> tuple[int, int, int, str]:
-    """Order DFlash sidecars in a repo listing / cache snapshot against the
-    weight actually being loaded.
-
-    dflash_preference_key ranks by precision and name alone, which is all a
-    single-model repo needs. A repo hosting more than one family also has to be
-    told which weight each sidecar belongs to, or ``dflash-model-A-Q8_0.gguf``
-    outranks the generic ``dflash-kquant.gguf`` on precision and model B is
-    launched with model A's drafter. Same rule the local scan applies in
-    detect_dflash_file, kept in one place so the download, the snapshot reuse
-    and the offline cache all pick the same file.
-
-    Three buckets: a sidecar naming this weight's family (most specific stem
-    first, as detect_mtp_file does), then one naming no weight present here,
-    then one naming a neighbour. The last is demoted rather than dropped, so a
-    repo whose only sidecar looks foreign still gets a fallback and today's
-    single-sidecar behaviour is unchanged.
-    """
-    precision, sort_name = dflash_preference_key(name)
-    if weight_name is not None and _drafter_matches_weight(name, weight_name, kind = "dflash"):
-        return 0, _drafter_stem_rank(name, kind = "dflash"), precision, sort_name
-    foreign = _drafter_names_other_weight(name, weight_name, other_weight_names)
-    return 2 if foreign else 1, 0, precision, sort_name
-
-
-def is_dflash_architecture(path: str) -> bool:
-    """Whether a GGUF really is a DFlash sidecar, decided by its header.
-
-    ``dflash-`` is a filename convention an ordinary weight can satisfy, by
-    accident or otherwise, and llama-server only discovers that at startup: it
-    refuses the file as ``--model-draft`` and the load falls back to no
-    speculation, after the bytes were already fetched. A DFlash sidecar declares
-    ``general.architecture = dflash``, which no real weight does, so that is what
-    settles it.
-
-    Kept here, beside the naming rules, because the local scan
-    (detect_dflash_file) and the download / cache reuse in llama_cpp all have to
-    apply it -- a remote path that trusted the prefix alone would download
-    gigabytes the launch then cannot use.
-    """
-    meta = read_gguf_general_metadata(str(path)) or {}
-    return (meta.get("general.architecture") or "").strip().lower() == "dflash"
-
+# Drafter naming, ranking and DFlash discovery live in utils.models.drafters, so
+# one rule serves the local scan, the download, the snapshot reuse and the
+# offline cache alike. Re-exported here deliberately: these names were part of
+# this module's surface and are imported from it across the backend and the
+# tests, so the move stays source compatible.
+from utils.models.drafters import (  # noqa: E402
+    _drafter_launch_path,
+    _drafter_matches_weight,
+    _drafter_names_other_weight,
+    _drafter_pairing_stem,
+    _drafter_split_is_complete,
+    _drafter_stem_rank,
+    _drafter_total_size,
+    detect_dflash_file,
+    dflash_precision_rank,
+    dflash_preference_key,
+    dflash_repo_preference_key,
+    dspark_precision_rank,
+    dspark_preference_key,
+    is_dflash_architecture,
+)
 
 def detect_mtp_file(
     path: str,
@@ -2122,162 +1951,6 @@ def detect_dspark_file(
         if accept is not None and not accept(launch):
             continue
         logger.info("Detected DSpark drafter: %s", launch)
-        return launch
-    return None
-
-
-def detect_dflash_file(
-    path: str,
-    search_root: Optional[str] = None,
-    accept: Optional[Callable[[str], bool]] = None,
-) -> Optional[str]:
-    """Find a DFlash sidecar for a local GGUF model.
-
-    Two things differ from detect_dspark_file, both forced by how DFlash is
-    published:
-
-    1. Root level only. ``dspark/`` is always a publisher's companion folder, so
-       that scan is safe; ``dflash/`` is a family name a user picks for real
-       weights (the reason llama_cpp._DRAFTER_DIR_KINDS leaves it out), so
-       reaching into it would launch a weight copy as --model-draft.
-    2. No filename pairing. The published sidecar is ``dflash-kquant.gguf``,
-       which names no model family at all, so _drafter_matches_weight would
-       reject the one file this exists to find. The header is checked instead:
-       a DFlash sidecar declares ``general.architecture = dflash``, which no
-       real weight does, and that is a stronger signal than a filename. It also
-       settles the adversarial case on its own, since a model merely CALLED
-       DFlash (``Qwen3.6-35B-A3B-DFlash-Q4_K_M.gguf``) reports its own
-       architecture.
-
-    A sidecar that does name a family (``dflash-Qwen3.6-27B-BF16.gguf``, the
-    scheme ggml-org uses) still wins over an unnamed one for the weight it
-    matches, so a multi-model folder attaches the specific sidecar first.
-
-    ``accept`` filters candidates in preference order, so a caller with extra
-    rules (a native lease) keeps scanning instead of treating the first
-    rejection as no sidecar at all.
-    """
-
-    def _rank(candidate: Path) -> tuple[int, int, int, int, str]:
-        # A sidecar naming THIS weight's family first, then any unpaired one,
-        # then precision, then total size so a split copy cannot outrank a
-        # smaller single file, then name for a stable order.
-        paired = _drafter_matches_weight(candidate.name, weight_name, kind = "dflash")
-        return (
-            0 if paired else 1,
-            _drafter_stem_rank(candidate.name, kind = "dflash") if paired else 0,
-            dflash_precision_rank(candidate.name),
-            _drafter_total_size(candidate),
-            candidate.name.lower(),
-        )
-
-    p = Path(path)
-    weight_name = p.name if p.suffix.lower() == ".gguf" else None
-    start_dir = p.parent if p.is_file() else p
-    dirs = [start_dir]
-    if search_root is not None:
-        dirs.append(Path(search_root))
-
-    candidates: list[Path] = []
-    other_weights: list[str] = []
-    seen: set[Path] = set()
-    # dict.fromkeys: search_root is the weight's own parent for a flat layout,
-    # and scanning it twice doubles the directory reads for nothing.
-    for root in dict.fromkeys(dirs):
-        try:
-            entries = list(root.iterdir())
-        except OSError:
-            continue
-        for candidate in entries:
-            lower = candidate.name.lower()
-            if not lower.endswith(".gguf"):
-                continue
-            # Prefix form only, deliberately. The shared companion predicates
-            # (_drafter_path_kind, is_mtp_drafter_path) know DFlash by the
-            # dflash- prefix, so accepting <model>-dflash.gguf here would let one
-            # file be a drafter for discovery AND a selectable Q8_0 main model in
-            # the quant picker, and choosing that variant would hand llama-server
-            # the drafter as the target. Teaching the predicate the suffix
-            # instead would hide a real model whose name merely ends in DFlash,
-            # which is the case #7811 exists to protect, so detection gives the
-            # form up rather than the picker giving up a model. No published
-            # DFlash sidecar uses it; the shipped one is dflash-kquant.gguf.
-            if not lower.startswith("dflash-"):
-                # Every other GGUF in the folder is a weight some sidecar could
-                # be naming. Recorded so a sidecar belonging to a NEIGHBOUR can
-                # be told apart from one naming no family at all (below).
-                other_weights.append(candidate.name)
-                continue
-            try:
-                # Collapse a split copy to shard 1 before ranking.
-                launch = _local_gguf_load_path(candidate)
-                # is_file() follows the link, so this also drops a dangling
-                # snapshot symlink and a directory named like a sidecar. Without
-                # it --model-draft gets a path llama-server cannot open, which
-                # fails the whole load rather than falling back to no
-                # speculation (detect_dspark_file guards the same way).
-                if not (launch.is_file() and _drafter_split_is_complete(launch)):
-                    continue
-                resolved = launch.resolve()
-            except OSError:
-                continue
-            if resolved in seen:
-                continue
-            seen.add(resolved)
-            candidates.append(launch)
-
-    # A sidecar naming a family that belongs to a NEIGHBOUR weight is that
-    # neighbour's drafter, not a generic one. _drafter_matches_weight is False
-    # both for it and for a sidecar naming no family (dflash-kquant.gguf), so
-    # ranking alone bucketed the two together and precision could float the
-    # foreign one to the top: loading model B beside dflash-model-A-Q8_0.gguf
-    # and dflash-kquant.gguf launched model A's drafter for model B. Both carry
-    # a real dflash header, so the architecture check behind the ranking cannot
-    # catch it. _drafter_names_other_weight decides against the weights actually
-    # present, which keeps the published unpaired sidecar eligible (its stem,
-    # "kquant", names no file here) without hardcoding which stems are precision
-    # tokens. Shared with the remote paths through dflash_repo_preference_key,
-    # so a download and a local scan agree on which sidecar belongs here.
-    if weight_name is not None and other_weights:
-        kept: list[Path] = []
-        for candidate in candidates:
-            if _drafter_names_other_weight(candidate.name, weight_name, other_weights):
-                logger.info(
-                    "detect_dflash_file: dropped %s (names another weight in this folder)",
-                    candidate.name,
-                )
-                continue
-            kept.append(candidate)
-        candidates = kept
-
-    for candidate in sorted(candidates, key = _rank):
-        # Resolve and validate before opening anything. A dflash-*.gguf in a
-        # directory reached through a native grant can be a symlink whose target
-        # sits outside the lease, and ``accept`` is what decides that; reading the
-        # header first opened the target before the answer arrived, which a later
-        # rejection cannot undo. Callers without a grant pass accept = None and
-        # see the same candidates in the same order as before.
-        try:
-            launch = _drafter_launch_path(candidate)
-        except OSError:
-            continue
-        if accept is not None and not accept(launch):
-            logger.info(
-                "detect_dflash_file: dropped %s (outside the granted directory)",
-                candidate.name,
-            )
-            continue
-        if not is_dflash_architecture(launch):
-            logger.info(
-                "detect_dflash_file: dropped %s (architecture %r is not dflash)",
-                candidate.name,
-                # Re-read only on the reject path, and header reads are cached by
-                # (path, mtime, size), so naming the offending architecture in the
-                # log costs nothing.
-                (read_gguf_general_metadata(launch) or {}).get("general.architecture"),
-            )
-            continue
-        logger.info("Detected DFlash drafter: %s", launch)
         return launch
     return None
